+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+HEAD
+----
+
+- triggering onSelect callback when using `set` method
+
 1.1.1
 -----